+++ lib_dsp/api/lib_dsp_transforms.h
@@ -42,12 +42,20 @@ lib_dsp_fft_complex_t;
 
 typedef struct
 {
-    short int re_1; // highest addr in memory
+    short int re_1;
     short int re_0;
     short int im_1;
-    short int im_0; // lowest addr in memory
+    short int im_0;
+}
+lib_dsp_fft_complex_twoshort_t;
+
+typedef struct
+{
+    short int re;
+    short int im;
 }
 lib_dsp_fft_complex_short_t;
+
 
 /** This function takes a the output from the function lib_dsp_fft_forward_complex and
  * reorders the data to form two frequency domain arrays.
@@ -83,18 +91,6 @@ void lib_dsp_fft_rebuild_one_real_input(lib_dsp_fft_complex_t pts[], unsigned N)
 
 void lib_dsp_fft_bit_reverse( lib_dsp_fft_complex_t pts[], int N );
 
-/** This function preforms index bit reversing on the the arrays around prior to computing an FFT. A
- * calling sequence for a forward FFT on two complex short int signals involves lib_dsp_fft_bit_reverse_short()
- * followed by lib_dsp_fft_forward_complex_short(), and for an inverse FFT it involves lib_dsp_fft_bit_reverse() followed
- * by lib_dsp_fft_inverse_complex_short(). In some cases bit reversal can be avoided, for example
- * when computing a convolution.
- *
- * \param pts   Array of complex input points
- *
- * \param N     number of points. Must be a power of 2, both re and im should be N long
- */
-void lib_dsp_fft_bit_reverse_short( lib_dsp_fft_complex_short_t pts[], int N );
-
 /** This function computes a forward FFT. The complex input signal is
  * supplied in an array of real and imaginary fixed-point values.
  * The same array is also used to store the output.
@@ -117,12 +113,25 @@ void lib_dsp_fft_forward_complex
     int                   N,
     const int             sine[] );
 
+/** This function preforms index bit reversing on the the arrays around prior to computing an FFT. A
+ * calling sequence for a forward FFT on two complex short int signals involves lib_dsp_fft_bit_reverse_twoshort()
+ * followed by lib_dsp_fft_forward_complex_twoshort(), and for an inverse FFT it involves lib_dsp_fft_bit_reverse_twoshort()
+ * followed by lib_dsp_fft_inverse_complex_twoshort(). In some cases bit reversal can be avoided, for example
+ * when computing a convolution.
+ *
+ * \param pts   Array of complex input points
+ *
+ * \param N     number of points. Must be a power of 2, both re and im should be N long
+ */
+void lib_dsp_fft_bit_reverse_twoshort( lib_dsp_fft_complex_twoshort_t pts[], int N );
+
+
 /** This function computes a forward FFT on two complex short int signals in one go.
  * The complex input signal is supplied in an array of real and imaginary fixed-point values.
  * The same array is also used to store the output.
  * The number of points must be a power of 2, and the array of sine values should contain a quarter sine-wave.
  * Use one of the lib_dsp_sine_N tables. The function does not perform bit reversal of the data.
- * If required then lib_dsp_fft_bit_reverse_short() should be called beforehand.
+ * If required then lib_dsp_fft_bit_reverse_twoshort() should be called beforehand.
  *
  * \param[in,out] pts   Array of complex input points. Must be 64-bit aligned to enable efficient double word loads and stores
  *
@@ -133,18 +142,54 @@ void lib_dsp_fft_forward_complex
  *                Arrays are provided in lib_dsp_tables.c; for example, for a 1024 point
  *                FFT use lib_dsp_sine_1024.
  */
+void lib_dsp_fft_forward_complex_twoshort(
+        lib_dsp_fft_complex_twoshort_t pts[],
+        int                         N,
+        const short                 sine[] );
+
+
+/** This function preforms index bit reversing on the the arrays around prior to computing an FFT. A
+ * calling sequence for a forward FFT on a complex short int signal involves lib_dsp_fft_bit_reverse_short()
+ * followed by lib_dsp_fft_forward_complex_short(), and for an inverse FFT it involves lib_dsp_fft_bit_reverse_short()
+ * followed by lib_dsp_fft_inverse_complex_short(). In some cases bit reversal can be avoided, for example
+ * when computing a convolution.
+ *
+ * \param pts   Array of complex input points
+ *
+ * \param N     number of points. Must be a power of 2, both re and im should be N long
+ */
+void lib_dsp_fft_bit_reverse_short( lib_dsp_fft_complex_short_t pts[], int N );
+
+/** This function computes a forward FFT on a complex short int signal.
+ * The complex input signal is supplied in an array of real and imaginary fixed-point values.
+ * The same array is also used to store the output.
+ * The number of points must be a power of 2, and the array of sine values should contain a quarter sine-wave.
+ * Use one of the lib_dsp_sine_N tables. The function does not perform bit reversal of the data.
+ * If required then lib_dsp_fft_bit_reverse_short() should be called beforehand.
+ *
+ * \param[in,out] pts   Array of complex input points.
+ *
+ * \param[in]     N     number of points. Must be a power of 2, both re and im should be N long
+ *
+ * \param[in]     sine  array of N/4+1 sine values, each represented as a sign bit,
+ *                and a 31 bit fraction. 1 should be represented as 0x7fffffff.
+ *                Arrays are provided in lib_dsp_tables.c; for example, for a 1024 point
+ *                FFT use lib_dsp_sine_1024.
+ */
 void lib_dsp_fft_forward_complex_short(
         lib_dsp_fft_complex_short_t pts[],
-        int                         N,
-        const short                 sine[] );
+        int N,
+        const short sine[] );
 
-
-void lib_dsp_fft_inverse_complex_short(
+/** Assembly version of lib_dsp_fft_forward_complex_short
+ * Not fully working yet
+ */
+void lib_dsp_fft_forward_complex_short_asm(
         lib_dsp_fft_complex_short_t pts[],
         int                         N,
         const short                 sine[] );
 
-/** This function computes the FFT of two real sequences in one go. It uses
+/** This function computes the FFT of two real signals in one go. It uses
  * a nifty trick (http://www.katjaas.nl/realFFT/realFFT.html) that enables
  * one to use a single complex FFT to compute two real FFTs simultaneously.
  * The real inputs should be in the first two real arrays, the output is in
@@ -170,6 +215,33 @@ void lib_dsp_fft_forward_tworeals
     lib_dsp_fft_complex_t two_im[],
     int       N,
     const int sine[] );
+
+/** This function computes the FFT of two real short int signals in one go. It uses
+ * a nifty trick (http://www.katjaas.nl/realFFT/realFFT.html) that enables
+ * one to use a single complex FFT to compute two real FFTs simultaneously.
+ * The real inputs should be in the first two real arrays, the output is in
+ * the real and imaginary arrays (the output of a real FFT is still a
+ * complex number).
+ *
+ * \param[in,out] two_re array containing two real short int signals on which to
+ *                       compute FFT, on output this array stores the real
+ *                       part of the complex FFT on the two real signals.
+ *
+ * \param[out]    two_im imaginary parts of complex FFT of two real signals in first array
+ *
+ * \param[in]     N      number of points
+ *
+ * \param[in]     sine  array of N/4+1 sine values, each represented as a sign bit,
+ *                and a 31 bit fraction. 1 should be represented as 0x7fffffff.
+ *                Arrays are provided in lib_dsp_tables.c; for example, for a 1024 point
+ *                FFT use lib_dsp_sine_1024.
+ */
+void lib_dsp_fft_forward_tworeals_short(
+        lib_dsp_fft_complex_short_t two_re[],
+        lib_dsp_fft_complex_short_t two_im[],
+        int N,
+        const short int sine[] );
+
 
 /** This function computes an inverse FFT. The complex input array is
  * supplied as two arrays of integers, with numbers represented as
@@ -220,6 +292,61 @@ void lib_dsp_fft_inverse_tworeals
     lib_dsp_fft_complex_t two_im[],
     int       N,
     const int sine[] );
+
+
+/** This function computes the inverse FFT on two sets of complex short int data that
+ * are known to result in real numbers only in one go. It uses a nifty
+ * trick (http://www.katjaas.nl/realFFT/realFFT.html) that enables one to
+ * use a single complex inverse FFT to compute two real inverse FFTs
+ * simultaneously. The outputs are in the two real arrays, the imaginary
+ * arrays are unchanged.
+ *
+ * \param[in,out] two_re array containing two real frequency domain signals on which to
+ *                       compute inverse FFT. On output this array stores two real
+ *                       time domain signals
+ *
+ * \param[out]    two_im array containing two imaginary frequency domain signals on which to
+ *                       compute inverse FFT
+ *
+ * \param[in]     N      number of points
+ *
+ * \param[in]     sine  array of N/4+1 sine values, each represented as a sign bit,
+ *                and a 31 bit fraction. 1 should be represented as 0x7fffffff.
+ *                Arrays are provided in lib_dsp_tables.c; for example, for a 1024 point
+ *                FFT use lib_dsp_sine_1024.
+ */
+void lib_dsp_fft_inverse_tworeals_short(
+        lib_dsp_fft_complex_short_t two_re[],
+        lib_dsp_fft_complex_short_t two_im[],
+        int N,
+        const short int sine[] );
+
+/** This function computes an inverse FFT. The complex input array is
+ * supplied as two arrays of integers, with numbers represented as
+ * fixed-point values. The number of points must be a power of 2, and the
+ * array of sine values should contain a quarter sine-wave. Use one of the
+ * lib_dsp_sine_N tables. The function does not perform a bit-twiddle -
+ * if required then lib_dsp_fft_bit_reverse() should be called beforehand.
+ *
+ * \param pts   Array of complex input points
+ *
+ * \param N     number of points. Must be a power of 2, both re and im should be N long
+ *
+ * \param sine  array of N/4+1 sine values, each represented as a sign bit,
+ *              and a 31 bit fraction. 1 should be represented as 0x7fffffff.
+ *              Arrays are provided in lib_dsp_tables.c; for example, for a 1024 point
+ *              FFT use lib_dsp_sine_1024.
+ */
+void lib_dsp_fft_inverse_complex_twoshort(
+        lib_dsp_fft_complex_twoshort_t pts[],
+        int                         N,
+        const short                 sine[] );
+
+
+void lib_dsp_fft_inverse_complex_short(
+        lib_dsp_fft_complex_short_t pts[],
+        int                         N,
+        const short                 sine[] );
 
 #endif