+++ tests/test_transforms.py
@@ -9,7 +9,7 @@ def runtest():
                                        'app_transforms', {})
      
     xmostest.run_on_simulator(resources['xsim'],
-                              '../AN00209_xCORE-200_DSP_Library/app_transforms/bin/tworeal_fft/app_transforms_tworeal_fft.xe',
+                              '../AN00209_xCORE-200_DSP_Library/app_transforms/bin/tworeals_fft/app_transforms_tworeals_fft.xe',
                               tester=tester)
 
 
@@ -26,12 +26,29 @@ def runtest():
 
     resources = xmostest.request_resource("xsim")
      
-    tester = xmostest.ComparisonTester(open('complex_short_int_fft_test.expect'),
+    tester = xmostest.ComparisonTester(open('app_fft_short_int_two_complex.xe.expect'),
                                        'lib_dsp', 'simple_tests',
                                        'app_fft_short_int', {})
      
     xmostest.run_on_simulator(resources['xsim'],
-                              '../AN00209_xCORE-200_DSP_Library/app_fft_short_int/bin/app_fft_short_int.xe',
+                              '../AN00209_xCORE-200_DSP_Library/app_fft_short_int/bin/two_complex/app_fft_short_int_two_complex.xe',
                               tester=tester)
 
+    resources = xmostest.request_resource("xsim")
+     
+    tester = xmostest.ComparisonTester(open('app_fft_short_int_complex.xe.expect'),
+                                       'lib_dsp', 'simple_tests',
+                                       'app_fft_short_int', {})
+     
+    xmostest.run_on_simulator(resources['xsim'],
+                              '../AN00209_xCORE-200_DSP_Library/app_fft_short_int/bin/complex/app_fft_short_int_complex.xe',
+                              tester=tester)
+
+    tester = xmostest.ComparisonTester(open('app_fft_short_int_tworeals.xe.expect'),
+                                       'lib_dsp', 'simple_tests',
+                                       'app_fft_short_int', {})
+     
+    xmostest.run_on_simulator(resources['xsim'],
+                              '../AN00209_xCORE-200_DSP_Library/app_fft_short_int/bin/tworeals/app_fft_short_int_tworeals.xe',
+                              tester=tester)